python_service/requirements.txt
flask==3.0.3
flask-cors==4.0.0
sqlalchemy>=2.0.36
flask-sqlalchemy>=3.1.1
psycopg2-binary==2.9.9
opencv-python-headless==4.10.0.84
numpy==2.1.3
requests==2.32.3
beautifulsoup4>=4.12.2
sentence-transformers>=2.2.2
google-generativeai>=0.1.0
python-dotenv>=1.0.0
jieba>=0.42.1
lightgbm>=3.3.5